cleaned up errors and added some refactoring
