+++ heatmap/Cargo.toml
@@ -17,7 +17,7 @@ crate-type = ["cdylib"]
 opt-level = "z"
 
 [dependencies]
-serde-wasm-bindgen = "0.4"
+serde-wasm-bindgen = "0.6.5"
 serde = { version = "1.0.213", features = ["derive"] }
 serde_derive = "1.0"
 wasm-bindgen = "0.2.91"
